% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accepted_object_classes.R
\name{accepted_object_classes}
\alias{accepted_object_classes}
\title{Accepted Object S3 Classes}
\usage{
accepted_object_classes()
}
\value{
character vector
}
\description{
Accepted Object S3 Classes
}
\examples{
accepted_object_classes()
}
