% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accepted_object_classes.R
\name{accepted_plot_classes}
\alias{accepted_plot_classes}
\title{Accepted Plot Classes}
\usage{
accepted_plot_classes()
}
\description{
Accepted Plot Classes
}
\examples{
accepted_plot_classes()
}
