% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accepted_object_classes.R
\name{accepted_table_classes}
\alias{accepted_table_classes}
\title{Accepted Table Classes}
\usage{
accepted_table_classes()
}
\description{
Accepted Table Classes
}
\examples{
accepted_table_classes()
}
