% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chunks.R
\name{create_chunks}
\alias{create_chunks}
\title{Create Code Chunks}
\usage{
create_chunks(length)
}
\arguments{
\item{length}{(\code{integer})\cr
integer specifying the number of code chunks to create.}
}
\value{
Code chunk character vector
}
\description{
Creates a series of R markdown code chunks that print the \code{x} object.
}
\examples{
create_chunks(2L) |>
  cat(sep = "\n")
}
