% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_html.R
\name{create_chunks_html}
\alias{create_chunks_html}
\title{Create Code Chunks for HTML output (internal)}
\usage{
create_chunks_html(length)
}
\arguments{
\item{length}{(\code{integer})\cr
integer specifying the number of code chunks to create.}
}
\description{
Creates R markdown chunks that render each \code{gt_tbl} object as HTML using
\code{gt::as_raw_html()}.
}
\keyword{internal}
