% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_yaml_header.R
\name{create_yaml_header}
\alias{create_yaml_header}
\title{Create YAML Header}
\usage{
create_yaml_header(object_path, pkg_to_attach, reference_docx)
}
\arguments{
\item{object_path}{(\code{path})\cr
path to saved RDS object file.}

\item{pkg_to_attach}{(\code{character})\cr
character vector of packages that are attached at the top of the R markdown
document. These are needed to access the appropraite print methods
for the loaded table object.}

\item{reference_docx}{(\code{path})\cr
path the reference Word document.}
}
\value{
YAML character vector
}
\description{
Creates the R markdown header, and first setup chunk.
The setup chunk quiets messages, attaches any needed packages,
and loads the object to be printed.
}
\examples{
create_yaml_header(
  object_path = tempfile(fileext = ".rds"),
  pkg_to_attach = "flextable",
  reference_docx = tempfile(fileext = ".docx")
) |>
  cat(sep = "\n")
}
