% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_txt.R
\name{create_yaml_header_txt}
\alias{create_yaml_header_txt}
\title{Create YAML Header for txt/markdown output (internal)}
\usage{
create_yaml_header_txt(object_path, pkg_to_attach)
}
\arguments{
\item{object_path}{(\code{path})\cr
path to saved RDS object file.}

\item{pkg_to_attach}{(\code{character})\cr
character vector of packages that are attached at the top of the R markdown
document. These are needed to access the appropriate print methods
for the loaded table object.}
}
\description{
Create YAML Header for txt/markdown output (internal)
}
\keyword{internal}
