% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_reference_docx.R
\name{get_reference_docx}
\alias{get_reference_docx}
\alias{get_all_reference_docx}
\title{Get Reference Document Path}
\usage{
get_reference_docx(orientation = "portrait", font = "Arial", font_size = 7)

get_all_reference_docx()
}
\arguments{
\item{orientation}{(\code{string})\cr
Must be one of "portrait", "landscape"}

\item{font}{(\code{string})\cr
Font of the page numbers.
Must be one of "Arial"}

\item{font_size}{(\code{integer})\cr
Font size of the page numbers.
Must be one of 7}
}
\value{
path
}
\description{
Get Reference Document Path
}
\examples{
get_reference_docx()
}
