% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-map.R
\name{map_unit}
\alias{map_unit}
\title{Map point estimates for units in a plot list.}
\usage{
map_unit(.pl, data, map, unit, value, group, ...)
}
\arguments{
\item{.pl}{plot list, generate with `plot_list` and pipe in.}

\item{data}{data for point estimate plots}

\item{map}{list of map df objects to merge data to}

\item{unit}{bare name of the unit column (in map and data)}

\item{value}{bare name of the value column.}

\item{group}{bare name of the variable to use to split plots on.}
}
\value{
plot list
}
\description{
For each plot in .pl, add a plot IFF the grouping factor
in 'data' has a matching level (when used with 'as.character').
}
