% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-core.R
\name{plot_list}
\alias{plot_list}
\title{Create a list of named empty plots.}
\usage{
plot_list(names)
}
\arguments{
\item{names, }{the names of each component plots in a vector.  These
will always be used with 'as.character'.}
}
\value{
a list of empty plots.
}
\description{
We often want to produce multi-page plots and ggplot offers
no straightfoward way to do this.  This is a minimal structure
to make that easy.
}
