% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-tpt.R
\name{plot_pt}
\alias{plot_pt}
\title{Add point estimates to a plot list.}
\usage{
plot_pt(.pl, data, time, value, group, ...)
}
\arguments{
\item{.pl}{plot list, generate with `plot_list` and pipe in.}

\item{data}{data for point estimate plots}

\item{time}{bare name of the time-axis column}

\item{value}{bare name of the value column.}

\item{group}{bare name of the variable to use to split plots on.}
}
\value{
plot list
}
\description{
For each plot in .pl, add a plot IFF the grouping factor
in 'data' has a matching level (when used with 'as.character').
}
