% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-tpt.R
\name{plot_pt_bounds}
\alias{plot_pt_bounds}
\title{Add point estimate bounds to a plot list}
\usage{
plot_pt_bounds(.pl, data, time, lb, ub, group, ...)
}
\arguments{
\item{.pl}{plot list, generate with `plot_list` and pipe in.}

\item{data}{data for time-series plots}

\item{time}{bare name of the time-axis column}

\item{lb}{bare name of the lower bound column.}

\item{ub}{bare name of the upper bound column.}

\item{group}{bare name of the variable to use to split plots on.}
}
\value{
plot list
}
\description{
For each plot in .pl, add a plot IFF the grouping factor
in 'data' has a matching level (when used with 'as.character').
}
