% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-core.R
\name{plot_titles}
\alias{plot_titles}
\title{Add titles to all plots}
\usage{
plot_titles(.pl, titles)
}
\arguments{
\item{.pl}{plot list, generate with `plot_list` and pipe in.}

\item{titles}{list, with names matching plot object names.}
}
\value{
plot list with titles added
}
\description{
For each plot in .pl, add a `ggtitle` with the provided strings.
}
