% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-unit.R
\name{plot_unit_bounds}
\alias{plot_unit_bounds}
\title{Add bounds estimates for units to a plot list.}
\usage{
plot_unit_bounds(.pl, data, unit, lb, ub, group, ...)
}
\arguments{
\item{.pl}{plot list, generate with `plot_list` and pipe in.}

\item{data}{data for point estimate plots}

\item{unit}{bare name of the unit-axis column}

\item{group}{bare name of the variable to use to split plots on.}

\item{value}{bare name of the value column.}
}
\value{
plot list
}
\description{
For each plot in .pl, add a plot IFF the grouping factor
in 'data' has a matching level (when used with 'as.character').
}
