% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_with_rmarkdown.R
\name{save_docx}
\alias{save_docx}
\title{Save as docx}
\usage{
save_docx(x, path, reference_docx = get_reference_docx("portrait"))
}
\arguments{
\item{x}{(\code{gtsummary}/\code{gt_tbl}/\code{flextable}/\code{ggplot}/\code{grob}/\code{list})\cr
object of class \code{'gtsummary'}, \code{'gt_tbl'} (gt table), \code{'flextable'}, \code{'ggplot'},
\code{'grob'}, or a list of these objects.}

\item{path}{(\code{path})\cr
path to save file to, e.g. "rendered_table.docx"}

\item{reference_docx}{(\code{path})\cr
path to reference document that when not \code{NULL} is passed to the
\verb{reference_docx:} R markdown field.}
}
\value{
(invisibly) a \code{string} corresponding to the content of the intermediate \code{.rmd} file that is rendered as
\code{.docx}.
}
\description{
Save input in a \code{.docx} file via R markdown.
}
\examples{
# create table
tbl <-
  cards::ADAE[1:150, ] |>
  gtsummary::tbl_hierarchical(
    variables = c(AESOC, AETERM),
    by = TRTA,
    denominator = cards::ADSL,
    id = USUBJID,
  )

# rendering to docx is slow, so these calls are not run during routine checks
\donttest{
# save as docx with flextable
gtsummary::as_flex_table(tbl) |>
  flextable::set_table_properties(layout = "autofit") |> # otherwise is going too wide
  save_docx(path = tempfile(fileext = ".docx"))

# Specify column width manually with flextable
gtsummary::as_flex_table(tbl) |>
  flextable::width(j = 1, width = 2) |>
  flextable::width(j = 2:3, width = 1.5) |>
  save_docx(path = tempfile(fileext = ".docx"))

# save as docx with gt
tbl |>
  gtsummary::as_gt() |>
  save_docx(path = tempfile(fileext = ".docx"))

# Specify column width manually with gt
tbl |>
  gtsummary::as_gt() |>
  gt::cols_width(
    dplyr::everything() ~ gt::px(100)
  ) |>
  save_docx(path = tempfile(fileext = ".docx"))

# split the table and save paginated table
gtsummary::tbl_split_by_rows(tbl, row_numbers = seq(20, nrow(tbl), by = 20)) |>
  save_docx(path = tempfile(fileext = ".docx"))

# save ggplot as docx
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()
save_docx(p, path = tempfile(fileext = ".docx"))
}

}
