% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_with_rmarkdown.R
\name{save_docx_with_rmarkdown}
\alias{save_docx_with_rmarkdown}
\title{Save as docx via R Markdown (internal)}
\usage{
save_docx_with_rmarkdown(
  x,
  path,
  reference_docx = get_reference_docx("portrait")
)
}
\arguments{
\item{x}{(\code{gtsummary}/\code{gt_tbl}/\code{flextable}/\code{ggplot}/\code{grob}/\code{list})\cr
object of class \code{'gtsummary'}, \code{'gt_tbl'} (gt table), \code{'flextable'}, \code{'ggplot'},
\code{'grob'}, or a list of these objects.}

\item{path}{(\code{path})\cr
path to save file to, e.g. "rendered_table.docx"}

\item{reference_docx}{(\code{path})\cr
path to reference document that when not \code{NULL} is passed to the
\verb{reference_docx:} R markdown field.}
}
\description{
Internal function that saves input in a \code{.docx} file via R markdown.
Use \code{\link[=save_docx]{save_docx()}} instead.
}
\keyword{internal}
