% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_txt.R
\name{save_txt_with_rmarkdown}
\alias{save_txt_with_rmarkdown}
\title{Save as txt (plain text / Markdown) via R Markdown (internal)}
\usage{
save_txt_with_rmarkdown(x, path, encoding = "UTF-8")
}
\arguments{
\item{x}{(\code{gt_tbl}/\code{gtsummary}/\code{list})\cr
object of class \code{'gt_tbl'} (gt table) or \code{'gtsummary'}, or a list of
such objects. Lists may contain a mix of \code{'gt_tbl'} and \code{'gtsummary'}
objects.}

\item{path}{(\code{path})\cr
path to save file to, e.g. \code{"rendered_table.txt"} or \code{"rendered_table.md"}.}

\item{encoding}{(\code{string})\cr
character encoding to use when writing the output file. Default is \code{"UTF-8"}.}
}
\description{
Internal function that saves table input as a plain text (Markdown) file
via R markdown. Use \code{\link[=save_txt]{save_txt()}} instead.
}
\keyword{internal}
