% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_with_rmarkdown.R
\name{save_with_rmarkdown}
\alias{save_with_rmarkdown}
\title{Save as docx}
\usage{
save_with_rmarkdown(x, path, reference_docx = get_reference_docx("portrait"))
}
\arguments{
\item{x}{(\code{gtsummary}/\code{gt_tbl}/\code{flextable}/\code{list})\cr
table object of class \code{'gtsummary'}, \code{'gt_tbl'} (gt table) or \code{'flextable'}, or a list
of table objects.}

\item{path}{(\code{path})\cr
path to save file to, e.g. "rendered_table.docx"}

\item{reference_docx}{(\code{path})\cr
Path to reference document that when not \code{NULL} is passed to the
\verb{reference_docx:} R markdown field.}

\item{output}{(\code{string})
Output format. For now, only \code{docx} output type is supported.}
}
\value{
x (invisibly)
}
\description{
Save Input via R markdown.
}
\examples{
# create table
tbl <-
  cards::ADAE[1:150,] |>
  gtsummary::tbl_hierarchical(
    variables = c(AESOC, AETERM),
    by = TRTA,
    denominator = cards::ADSL,
    id = USUBJID,
  )

# save as docx
gtsummary::as_flex_table(tbl) |>
  save_with_rmarkdown(path = tempfile(fileext = ".docx"))

# split the tqble and save paginatted table
gtsummary::tbl_split_by_rows(tbl, row_numbers = seq(20, nrow(tbl), by = 20)) |>
  save_with_rmarkdown(path = tempfile(fileext = ".docx"))
}
