% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accepted_classes.R
\name{accepted_object_classes}
\alias{accepted_object_classes}
\alias{accepted_table_classes}
\alias{accepted_plot_classes}
\title{Accepted Classes}
\usage{
accepted_table_classes()

accepted_plot_classes()
}
\value{
character vector
}
\description{
Return the acceptable classes for different types of output.
}
\examples{
accepted_table_classes()
accepted_plot_classes()
}
