% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accepted_table_classes.R
\name{accepted_table_classes}
\alias{accepted_table_classes}
\title{Accepted Table S3 Classes}
\usage{
accepted_table_classes()
}
\value{
character vector
}
\description{
Accepted Table S3 Classes
}
\examples{
accepted_table_classes()
}
