% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pager-package.R
\docType{package}
\name{pager-package}
\alias{pager}
\alias{pager-package}
\title{pager: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Authors:
\itemize{
  \item Benoit Falquet \email{benoit.falquet@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
