% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_txt.R
\name{save_txt}
\alias{save_txt}
\title{Save as txt (plain text / Markdown)}
\usage{
save_txt(x, path, encoding = "UTF-8")
}
\arguments{
\item{x}{(\code{gt_tbl}/\code{list})\cr
object of class \code{'gt_tbl'} (gt table), or a list of \code{'gt_tbl'} objects.}

\item{path}{(\code{path})\cr
path to save file to, e.g. \code{"rendered_table.txt"} or \code{"rendered_table.md"}.}

\item{encoding}{(\code{string})\cr
character encoding to use when writing the output file. Default is \code{"UTF-8"}.}
}
\value{
(invisibly) a \code{string} corresponding to the content of the intermediate
\code{.rmd} file that is rendered.
}
\description{
Save a \code{gt_tbl} object as a plain text (Markdown-formatted) file via R markdown.
Only \code{gt_tbl} objects (from the gt package) are supported for plain text output.
}
\examples{
# create gt table
tbl <- gt::gt(head(mtcars))

# save as txt
tbl |>
  save_txt(path = tempfile(fileext = ".txt"))

# save a list of gt tables as txt
list(gt::gt(head(mtcars)), gt::gt(tail(mtcars))) |>
  save_txt(path = tempfile(fileext = ".txt"))

}
