% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_txt.R
\name{save_txt}
\alias{save_txt}
\title{Save as txt (plain text / Markdown)}
\usage{
save_txt(x, path, encoding = "UTF-8")
}
\arguments{
\item{x}{(\code{gtsummary}/\code{gt_tbl}/\code{flextable}/\code{list})\cr
object of class \code{'gtsummary'}, \code{'gt_tbl'} (gt table), \code{'flextable'},
or a list of these objects.}

\item{path}{(\code{path})\cr
path to save file to, e.g. \code{"rendered_table.txt"} or \code{"rendered_table.md"}.}

\item{encoding}{(\code{string})\cr
character encoding to use when writing the output file. Default is \code{"UTF-8"}.}
}
\value{
(invisibly) a \code{string} corresponding to the content of the intermediate
\code{.rmd} file that is rendered.
}
\description{
Save table input as a plain text (Markdown-formatted) file via R markdown.
Plots are not supported for plain text output.
}
\examples{
# create table
tbl <-
  cards::ADAE[1:150, ] |>
  gtsummary::tbl_hierarchical(
    variables = c(AESOC, AETERM),
    by = TRTA,
    denominator = cards::ADSL,
    id = USUBJID,
  )

# save as txt with gtsummary
tbl |>
  save_txt(path = tempfile(fileext = ".txt"))

# save as txt with flextable
gtsummary::as_flex_table(tbl) |>
  save_txt(path = tempfile(fileext = ".txt"))

# save as txt with gt
tbl |>
  gtsummary::as_gt() |>
  save_txt(path = tempfile(fileext = ".txt"))

# save a paginated table as txt
gtsummary::tbl_split_by_rows(tbl, row_numbers = seq(20, nrow(tbl), by = 20)) |>
  save_txt(path = tempfile(fileext = ".txt"))

}
